pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sim-as-a-judge"
version = "0.1.0"
description = "Physics-grounded validation framework for synthetic robot trajectory data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "datasets>=2.18.0",
    "opencv-python-headless>=4.8.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
include = ["sim_judge*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
